test: enable subset simulation toy model test
