+++ matflow/data/workflows/subset_simulation_toy_model.yaml
@@ -13,7 +13,6 @@ loops:
 resources:
   any:
     combine_scripts: true
-    random_seed: 1234
 
 tasks:
   - schema: sample_direct_MC
